Give the framework the platform work its applications were writing
The audit of PLAN.md found eleven items ticked whose claims the code did not
support. This closes them, and where something genuinely cannot be done on a
platform it now says so through a capability rather than registering a backend
that can only fail.

Platform backends the framework was missing

  Desktop had no power monitor at all, so an application pacing itself by heat
  read every Mac as unsupported and could never throttle. `desktop_power.rs`
  reads thermal pressure from `NSProcessInfo` — Foundation, not UIKit, so a Mac
  and an iPhone answer the same question from the same call, and
  `apple_thermal.rs` is the one mapping both shells use. Linux reports battery
  from `/sys/class/power_supply` and answers `Unsupported` for thermal: the
  kernel exposes temperatures, not pressure, and a threshold this module
  invented would read as a measurement while being a guess.

  Desktop and iOS had no bundled-asset reader. Both now stream: a desktop
  executable's resources are found beside it, in a `.app`'s `Resources`, or in
  the working directory, and an asset name that climbs out of its root is
  refused before it becomes an arbitrary file read.

  Files dropped on a desktop window and documents named on the command line now
  arrive as the same `IncomingContent` a share does. A macOS `.app` opened
  through the Finder is told by an Apple Event that winit does not surface, so
  drops work and opens do not — said here rather than implied away.

  Resident memory on Apple targets comes from `task_info`; Darwin has no
  `/proc` to read a process's own footprint from.

One streaming reader, not four

  The desktop, iOS and Android backends each carried the same chunk-filling
  loop, and the crate's own fallback carried a fourth. `StreamingAssetReader`
  is now the only one: a backend supplies a stream and, when its asset is a
  span of something larger, how many bytes are its own. The Android copy had
  been naming the empty string in its read errors.

Update discovery split from update installation

  `AppUpdateCapabilities` separates `check` from `install`, because a platform
  can have one without the other. Android does both. Desktop and iOS discover a
  release through `GitHubAppUpdater` — written once against the framework's own
  `HttpClient` — and report `install: false`: App Store Review Guideline 3.3.2
  forbids an iOS application from replacing its own binary, and the framework
  owns no desktop installer. The trait defaults to claiming neither, so a
  backend that forgets to declare is refused rather than allowed to fail at the
  platform boundary, and the checker reports `check: false` when the build has
  no HTTP client to make the request with.

  Every way those two entry points can fail now reaches the observable status,
  not only the returned `Result`. Four early returns — no updater registered,
  or a backend that does not claim that half — published nothing, so a screen
  watching the update state saw no change at all when it asked a host that
  could not answer. A consumer application had grown a workaround for exactly
  that, mirroring the `Result` into the status by hand; the workaround is gone
  and a test holds the behaviour.

A dialog now traps focus

  A text field behind an open dialog could take focus from it, leaving the
  platform keyboard attached to a field the user could not see. A field carries
  the modal depth it was composed at, and a focus request shallower than the
  open depth is refused; closing a dialog releases the focus it held. The
  enforcement path reads the modal stack's own count rather than the observable
  mirror, which belongs to whichever runtime first allocated it — reading that
  from a pointer callback panicked when the runtime was another one.

The coverage measure was counting production code as test code

  `test_corpus` took each file from its first `#[cfg(test)]` to the end. That
  attribute also introduces test-only imports, and 138 files carry one near the
  top — `render.rs` at 0.1% of the file, handing 971KB of production source to
  the "test corpus". Every public function below such a marker certified itself
  as tested by its own source position. The corpus is now the braced body of
  each `#[cfg(test)] mod`, and the figure is 3031/3396 (89.3%) with 365
  untested, not 2628/2689 (97.7%) with 61. PLAN.md records that as the gap it
  is; a test written to raise a number tests the implementation it was written
  against.

Also

  The media backend gained its own decoder and output device in an earlier
  commit; a seek there played from up to a packet early and is fixed with a
  test that catches it. Clippy is at zero on Linux, macOS, the iOS target and
  the wasm target — twenty-eight sites the previous toolchain did not flag. The
  HTTP user agent came from a literal `cranpose/0.1`; it comes from the crate
  version now, and GitHub's API refuses a request carrying none. The workspace
  has one answer to where a test writes real files, behind
  `cranpose-core/test-helpers`, instead of a copy per crate and two tests
  writing to tmpfs.

+++ crates/cranpose-media/src/sink.rs
@@ -610,7 +610,7 @@ mod tests {
             produced += 1;
         }
         assert!(
-            (produced as i32 - (frames as i32 / 2)).abs() <= 2,
+            (produced - (frames as i32 / 2)).abs() <= 2,
             "produced {produced} from {frames}"
         );
     }
@@ -626,7 +626,7 @@ mod tests {
             produced += 1;
         }
         assert!(
-            (produced as i32 - (frames as i32 / 2)).abs() <= 2,
+            (produced - (frames as i32 / 2)).abs() <= 2,
             "produced {produced} from {frames}"
         );
     }
